[css-nesting] Reject invalid nested selectors from CSSOM.
Using CSSOM, we could add nested selectors that would be unparsable
(those that start with a tag selector); plug that hole and add tests
for it. Note that this probably also fixes that we wouldn't set
the parent rule correctly in these cases.

See https://github.com/w3c/csswg-drafts/issues/7971.

Bug: 1095675
Change-Id: Ieedde3a2b07873d940dcd0a3f0aa68159259b729