Convert animation-timeline property value to tree-scoped
Following the plan in crrev.com/c/4167268, this patch converts the
animation-timeline property to use tree-scoped-populated CSS values
instead of ScopedCSSValue.

There's a minor behavior change that the support of <string> values in
the property has been removed. This is because:
- The spec doesn't allow <string> values [1]
- Allowing <string> values will complicate the implementation, as it
  requires populating TreeScope into a CSSStringValue, which doesn't
  make much sense

Note: animation-timeline is currently an experimental feature, so we
don't need any Blink intents to modify its behavior.

[1] https://drafts.csswg.org/css-animations-2/#propdef-animation-timeline

Bug: 1395026
Change-Id: I8a44e8d2759a49368cbe864d8c3285aa0a79dca3